Move some things out of the test into the library.
